Command

Description

Module in charge of parsing the CLI command and constructing a domain specific representation of the instruction to execute.

Documentation

data GenCommand

GenCommand represents the instructions the program will execute. It contains what we are going to copy, its name and content substitutions as well as whether or not to treat the artifact as a module.

mkGenCommand

Arguments

:: String
What do we want to copy, or which template do we want to use.

-> String
The name that will be given to the new files.

-> Map String String
Mapping of values to substitute in the templates content.

-> Bool
Should the output files be treated as a module. If it is a module it will create a directory with the name and put the files inside.

-> GenCommand

GenCommand factory

parserOptions :: ParserInfo GenCommand

This is the configuration given to Opt-Parser Applicative. This is what will be used in the cli program, the messages and banners displayed. It is a parser for GenCommand.