Config

Description

Contains metadata for the application as well as values needed for Progen to run.

Documentation

data Dotfile

It represents the data coming from the configuration file. More data can be needed but may not be configurable. The Dotfile is used to create a GenConfig which will be used by the application.

data ConfigError

Encompasses all possible errors in the Config module

handleConfigResult :: Maybe GenConfig -> Either ConfigError GenConfig

Transform maybe config to an Either of the correct Domain error or the valid Config

data GenConfig

This is the configuration used by the application wtih all the metadata needed to generate code.

dotfileName :: String

Name for the dotfile where the configuration will be read from

mkDotfile

Arguments

:: Last AbsDir
Absolute directory path pointing to the root of the project using Progen.

-> Last AbsDir
Where are the templates. At the moment it can be out side of the root.

-> Last (Map String RelDir)
Configuration for the output. Keys are matched on the names of the templates, the values are relative directory paths inside the root.

-> Last Char
Character used to separate the template filename and make use of suffix.

-> Dotfile

Factory function for constructing a Dotfile

mkConfig :: Dotfile -> String -> Maybe GenConfig

Factory to make a GenConfig from a JSONString. It will attempt to construct a Dotfile from its values and then combine it with a default config. The result will be the GenConfig that will be returned.

>>> mkConfig "{ \"root\": \"/dummy/project\", \"templates\": \"/dummy/project/.progen/templates\", \"filenameSeparator\": \".\", \"output\": { \"component\": \"./components\" }}"