Template

Description

Templates represent the file to be written while holding information from the source template.

Documentation

data Template

A Template represents the file that is going to be written. It contains the content from the source template as well as the source path which is the filename of the source template

data TemplateError

Encompasses all possible errors in the Template module

mkTemplate :: String -> String -> String -> String -> String -> Template

Template factory

getTemplateFiles :: AbsDir -> GenConfig -> GenCommand -> IO (Either TemplateError [Template])

Iterates through all files in Progen template's source directory. For each file in the directory it will match all files that start with the what passed in the command.