Transformations

Description

Transformations are applied to the content of a template where each instance of a key found in the Commands substitutions Map is searched in the template content and replaced with the substitution value.

Documentation

transformContent :: GenCommand -> Template -> Template

Apply substitutions to templates content