Writer

Description

Persist template. Note: If there are many templates the operation is not atomic.

Documentation

write

Arguments

:: Bool
Whether to write the file as a module. If true, it will create a directory and save the templates in it.

-> GenConfig
Config

-> Template
Template to write to the output directory

-> IO (Either String String)

Write template file to output directory.