Config

Description

Contains metadata for the application as well as values needed for Progen to run.

Documentation

data GenConfig

This is the configuration used by the application wtih all the metadata needed to generate code.

dotfileName :: String

Name for the dotfile where the configuration will be read from

mkConfig :: String -> Maybe GenConfig

Factory to make a GenConfig from a JSONString. It will attempt to construct a Dotfile from its values and then combine it with a default config. The result will be the GenConfig that will be returned.

>>> mkConfig "{ \"root\": \"/dummy/project\", \"templates\": \"/dummy/project/.progen/templates\", \"filenameSeparator\": \".\", \"output\": { \"component\": \"./components\" }}"